Fixed coverity findings

+++ src/Outputs.h
@@ -39,7 +39,7 @@ void setOutputs(unsigned long vlvBits);
 
 void processAutoValve();
 
-unsigned long getValveBits();
+uint32_t getValveBits();
 #endif //#ifdef PVOUT
 
 unsigned long computeValveBits();
